Split serialize test over 3 files
